Add command struct + more (read desc)
Big commit and does not work, sorry.

This contains a lot of changes to the structure of how the shell will be programmed.
I haven't thought of a better way to write this, or design it rather, so it is
reading/writing memory expensive. It does however set us up for the future stages,
as we are lucky to have a static specification, it does help to build
scaffolding to the program.

Furthermore, added a "constants.h" file, used for only macros, previously placed
macros in other headers can be found inside this file.

Note: A lot of the functions are undocumented, this is something I will fix later.
At this point in time, please do message the group and dicuss whether you like or
disagree with the structure. I frankly do not like for a couple of reasons, but it
does improve the cost of time that it would take to implement it efficiently.

Simply put, the command struct contains more than it needs depending on it's type.
The type is defined as 3 enums, BUILTIN, EXECUTE AND ALIAS. Futureproof.

This allows us to expand an API as needed, and the command will flex to reflect what
is required of its type.